# adapter/inbound/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Blog server.",
        "title": "Blog Service Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "paths": {
        "/blogs/{id}": {
            "get": {
                "description": "get blog by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a blog",
                "operationId": "get-blog-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Blog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.blogResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.blogResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "a rich body for blog",
                    "type": "string",
                    "example": "body"
                },
                "id": {
                    "description": "blog identify",
                    "type": "integer",
                    "example": 10086
                },
                "title": {
                    "description": "title of blog",
                    "type": "string",
                    "example": "title"
                }
            }
        }
    }
}